delete CommentLinkButton and add it in ContentLinkButtons instead
